requirements.txt
socrata-py==0.4.0
pandas
prettytable
termcolor
